PolygonMarker uses BitmapMixin now, boundary of canvas is now double instead of int

+++ mapsforge_flutter/lib/src/graphics/mapcanvas.dart
@@ -65,7 +65,7 @@ abstract class MapCanvas {
   //
   // void setAntiAlias(bool aa);
   //
-  void setClip(int left, int top, int width, int height);
+  void setClip(double left, double top, double width, double height);
 
   // void setClipDifference(int left, int top, int width, int height);
   //
